Add package for simplified JWT usage (#17)
This package will be shared between the auth service and proxies that must
authenticate requests before delegating to the downstream service or handler.

This package standardizes on using the RSA-256 signing method. The auth service
will create and sign tokens using the private key and any consumer which requires
authenticating a token will have a public key used to verify the token.

Note that JWTs trade-off revocation for decentralization. In the use for Arcus
revocation will be an uncommon need, however request verification must occur
all the time. The decentralization property allows for local verification using
the public key without requiring a network call.

Signed-off-by: Byron Ruth <b@devel.io>